LayoutTests/fast/events/ios/rotation/basic-rotation-expected.txt

Before rotation
PASS window.innerWidth is 390
PASS window.innerHeight is 797
In resize event handler:
PASS window.innerWidth is 844
PASS window.innerHeight is 390

In orientationchange event handler:
PASS window.innerWidth is 844
PASS window.innerHeight is 390

After rotation
PASS window.innerWidth is 844
PASS window.innerHeight is 390
Rotation test.
